create interception driver wrapper
Not well tested or anything but easy_driver.py has the EasyDriver class which is where I moved all the previous interception driver functions to.

It will initialize the driver, and find a valid keyboard/mouse. It has functions to move the mouse and right click and a function to press a key.